Factor out common AddFileRequest, RemoveFileRequest logic
